+++ lib/action_client/engine.rb
@@ -1,5 +1,7 @@
 module ActionClient
   class Engine < ::Rails::Engine
+    config.action_client = ActiveSupport::OrderedOptions.new
+
     initializer "action_client.dependencies" do |app|
       ActionClient::Base.append_view_path app.paths["app/views"]
       ActionClient::Base.config_path = Pathname(app.paths["config"].first)
@@ -7,7 +9,7 @@ module ActionClient
 
     initializer "action_client.middleware" do
       ActionClient::Base.middleware = ActionDispatch::MiddlewareStack.new do |stack|
-        stack.use ActionClient::Middleware::ResponseParser
+        stack.use ActionClient::Middleware::Parser, config.action_client
         stack.use Rack::ContentLength
         stack.use Rails::Rack::Logger, [ActionClient::Middleware::Tagger]
       end