+++ lib/action_client/parser_error.rb
@@ -1,0 +1,3 @@
+module ActionClient
+  ParserError = ParseError
+end