+++ test/integration_test_case.rb
@@ -22,5 +22,14 @@ module ActionClient
     setup do
       ActionClient::Base.defaults = ActiveSupport::OrderedOptions.new
     end
+
+    def override_configuration(configuration, &block)
+      originals = configuration.dup
+
+      yield(configuration)
+    ensure
+      originals.each { |key, value| configuration[key] = value }
+      configuration.delete_if { |key, _| originals.keys.exclude?(key) }
+    end
   end
 end